+++ cumulus/preflighttaskpage.cpp
@@ -747,8 +747,11 @@ bool PreFlightTaskPage::createFlarmTaskList( FlightTask* flightTask )
   QString fn = GeneralConfig::instance()->getUserDataDirectory() + "/cumulus-flarm.tsk";
 
   // Save one backup copy.
-  QFile::remove( fn + ".bak" );
-  QFile::rename( fn, fn + ".bak" );
+  if( QFileInfo(fn).exists() )
+    {
+      QFile::remove( fn + ".bak" );
+      QFile::rename( fn, fn + ".bak" );
+    }
 
   QFile f(fn);
